Add some intelligence to the computer
Added a function that, in case that player1 is first to play,
it checks if the card played is worth some points.
If the function returns true, the computer tries to play a higher
rank of the same suit.
If it doesn't have it, it tries to play a "briscola".
If it also doesn't have a "briscola" it plays a card worth no points